+++ Debug/Types/Data.lua
@@ -11,7 +11,7 @@
 ---@field showRealms boolean
 ---@field showGuildInformation boolean
 ---@field currentCharacterMarker AE_CurrentCharacterMarkerId
----@field announceKeystones { autoParty: boolean, autoGuild: boolean, multiline: boolean, multilineNames: boolean}
+---@field announceKeystones { autoParty: boolean, multiline: boolean, multilineNames: boolean}
 ---@field announceResets boolean
 ---@field vault { raids: boolean, dungeons: boolean, world: boolean }
 ---@field prey { enabled: boolean, hiddenDifficulties: table<number, boolean> }